added people to handlers
