+++ routes.go
@@ -84,13 +84,13 @@ func NewRouter() *http.Server {
 
 		r.Get("/ask", ask)
 		r.Get("/poll/{challenge}", poll)
-		r.Get("/person/{pubkey}", getPersonByPubkey)
-		r.Get("/person/uuid/{uuid}", getPersonByUuid)
-		r.Get("/person/uuid/{uuid}/assets", getPersonAssetsByUuid)
-		r.Get("/person/githubname/{github}", getPersonByGithubName)
+		r.Get("/person/{pubkey}", handlers.GetPersonByPubkey)
+		r.Get("/person/uuid/{uuid}", handlers.GetPersonByUuid)
+		r.Get("/person/uuid/{uuid}/assets", handlers.GetPersonAssetsByUuid)
+		r.Get("/person/githubname/{github}", handlers.GetPersonByGithubName)
 
-		r.Get("/github_issue/{owner}/{repo}/{issue}", getGithubIssue)
-		r.Get("/github_issue/status/open", getOpenGithubIssues)
+		r.Get("/github_issue/{owner}/{repo}/{issue}", handlers.GetGithubIssue)
+		r.Get("/github_issue/status/open", handlers.GetOpenGithubIssues)
 		r.Post("/save", postSave)
 		r.Get("/save/{key}", pollSave)
 	})
@@ -106,13 +106,13 @@ func NewRouter() *http.Server {
 		r.Put("/tribeactivity/{uuid}", putTribeActivity)
 		r.Put("/tribepreview/{uuid}", setTribePreview)
 		r.Delete("/bot/{uuid}", handlers.DeleteBot)
-		r.Post("/person", createOrEditPerson)
+		r.Post("/person", handlers.CreateOrEditPerson)
 		r.Post("/verify/{challenge}", verify)
-		r.Post("/badges", addOrRemoveBadge)
+		r.Post("/badges", handlers.AddOrRemoveBadge)
 		r.Put("/bot", handlers.CreateOrEditBot)
-		r.Delete("/person/{id}", deletePerson)
+		r.Delete("/person/{id}", handlers.DeletePerson)
 		r.Delete("/channel/{id}", deleteChannel)
-		r.Delete("/ticket/{pubKey}/{created}", deleteTicketByAdmin)
+		r.Delete("/ticket/{pubKey}/{created}", handlers.DeleteTicketByAdmin)
 	})
 
 	r.Group(func(r chi.Router) {
